reorganize files and folder
