+++ src/component/navbar/Navbar.css
@@ -1,0 +1,8 @@
+nav {
+  display: flex;
+  justify-content: center;
+}
+
+a {
+  margin-left: 10px;
+}
