--- src/container/BlockedList.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function BlockedList() {
-  return <h1>Blocked list</h1>;
-}
