--- src/container/Home.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Home() {
-  return <h1>Home</h1>;
-}
