+++ src/container/blockedList/BlockedList.js
@@ -1,0 +1,11 @@
+import React from "react";
+import BlockList from "../../component/blockList/BlockList";
+
+export default function BlockedList() {
+  return (
+    <>
+      <h1>Blocked list page</h1>
+      <BlockList />
+    </>
+  );
+}
