+++ src/container/home/Home.css
@@ -1,0 +1,6 @@
+.home {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
