+++ src/container/home/Home.js
@@ -1,0 +1,11 @@
+import React from "react";
+import "./Home.css";
+
+export default function Home() {
+  return (
+    <div className="home">
+      <h1>Our Top Rated Movies List</h1>
+      <h2>This is the home page</h2>
+    </div>
+  );
+}
