+++ src/layout/index.js
@@ -1,9 +1,9 @@
 import { Routes, Route } from "react-router-dom";
-import Home from "../container/Home";
-import Navbar from "../component/Navbar";
-import MovieList from "../container/MovieList";
-import LikedList from "../container/LikedList";
-import BlockedList from "../container/BlockedList";
+import Home from "../container/home/Home";
+import Navbar from "../component/navbar/Navbar";
+import MovieList from "../container/movieList/MovieList";
+import LikedList from "../container/likedList/LikedList";
+import BlockedList from "../container/blockedList/BlockedList";
 
 export default function Layout() {
   return (
